Changed to While true
System should exit when user select menu 5. older version had while menu is between 0 and 5 to run the code. but when Menu was = to or greater then 5, it would exit. i now just add a While (True) instead to keep the while running but if the user put a value greater than 5 to just print a message to ask to enter a value between 1 and 5.